DFE-1175 - new skeleton Create Release page, and refactoring edit release code to reuse in Create Release process
