+++ src/explore-education-statistics-admin/src/pages/admin-dashboard/components/DashboardReleaseSummary.tsx
@@ -14,7 +14,7 @@ import React, { useContext } from 'react';
 import { format } from 'date-fns';
 import Details from '@common/components/Details';
 import { LoginContext } from '@admin/components/Login';
-import { setupRoute } from '../../../routes/edit-release/routes';
+import { setupRoute } from '@admin/routes/edit-release/routes';
 
 const getLiveLatestLabel = (isLive: boolean, isLatest: boolean) => {
   if (isLive && isLatest) {
