+++ src/explore-education-statistics-admin/src/pages/release/create-release/CreateReleasePage.tsx
@@ -1,0 +1,20 @@
+import Page from "@admin/components/Page";
+import React from 'react';
+
+const CreateReleasePage = () => {
+
+  return (
+    <Page
+      wide
+      breadcrumbs={[
+        {
+          name: 'Create new release',
+        },
+      ]}
+    >
+      hello
+    </Page>
+  );
+};
+
+export default CreateReleasePage;
