+++ src/explore-education-statistics-admin/src/pages/release/edit-release/data/ReleaseDataPage.tsx
@@ -1,12 +1,12 @@
-import ReleasePageTemplate from '@admin/pages/edit-release/components/ReleasePageTemplate';
-import ReleaseDataUploadsSection from '@admin/pages/edit-release/data/ReleaseDataUploadsSection';
-import ReleaseFileUploadsSection from '@admin/pages/edit-release/data/ReleaseFileUploadsSection';
 import service from '@admin/services/common/service';
-import { IdTitlePair } from '@admin/services/common/types';
+import {IdTitlePair} from '@admin/services/common/types';
 import Tabs from '@common/components/Tabs';
 import TabsSection from '@common/components/TabsSection';
-import React, { useEffect, useState } from 'react';
-import { RouteComponentProps } from 'react-router';
+import React, {useEffect, useState} from 'react';
+import {RouteComponentProps} from 'react-router';
+import ReleasePageTemplate from '../components/ReleasePageTemplate';
+import ReleaseDataUploadsSection from './ReleaseDataUploadsSection';
+import ReleaseFileUploadsSection from './ReleaseFileUploadsSection';
 
 interface MatchProps {
   releaseId: string;
