+++ src/explore-education-statistics-admin/src/routes/edit-publication/routes.ts
@@ -1,6 +1,6 @@
 export default {
   createPublication: {
-    route: '/topic/:topicId/publication/create',
-    generateLink: (topicId: string) => `/topic/${topicId}/publication/create`,
+    route: '/topic/:topicId/create-publication',
+    generateLink: (topicId: string) => `/topic/${topicId}/create-publication`,
   }
 }
