+++ src/explore-education-statistics-admin/src/routes/edit-release/routes.ts
@@ -1,10 +1,10 @@
-import ReleaseSetupEditPage from '../../pages/edit-release/ReleaseSetupEditPage';
-import ReleaseSetupPage from '../../pages/edit-release/ReleaseSetupPage';
-import ReleaseDataPage from '../../pages/edit-release/data/ReleaseDataPage';
-import ReleaseBuildTablesPage from '../../pages/edit-release/ReleaseBuildTablesPage';
-import ReleaseTablesPage from '../../pages/edit-release/ReleaseTablesPage';
-import ReleaseContentPage from '../../pages/edit-release/ReleaseContentPage';
-import ReleasePublishStatusPage from '../../pages/edit-release/ReleasePublishStatusPage';
+import ReleaseSetupEditPage from '@admin/pages/release/edit-release/setup/ReleaseSetupEditPage';
+import ReleaseSetupPage from '@admin/pages/release/edit-release/setup/ReleaseSetupPage';
+import ReleaseDataPage from '@admin/pages/release/edit-release/data/ReleaseDataPage';
+import ReleaseBuildTablesPage from '@admin/pages/release/edit-release/ReleaseBuildTablesPage';
+import ReleaseTablesPage from '@admin/pages/release/edit-release/ReleaseTablesPage';
+import ReleaseContentPage from '@admin/pages/release/edit-release/ReleaseContentPage';
+import ReleasePublishStatusPage from '@admin/pages/release/edit-release/ReleasePublishStatusPage';
 
 export interface ReleaseRoute {
   path: string;
@@ -123,4 +123,13 @@ export const editRoutes = [
   publishStatusEditRoute,
 ];
 
-export default [...viewRoutes, ...editRoutes];
+export default {
+  manageReleaseRoutes: [
+    ...viewRoutes,
+    ...editRoutes,
+  ],
+  createReleaseRoute: {
+    route: '/publication/:publicationId/create-release',
+    generateLink: (publicationId: string) => `/publication/${publicationId}/create-release`
+  },
+};
